Add labels and taints to the list machinepools command
Related: SDA-8219
